Rename library to bytebuild. Use Data.Bytes namespace.
