+++ bytebuild.cabal
@@ -1,5 +1,5 @@
 cabal-version: 2.2
-name: small-bytearray-builder
+name: bytebuild
 version: 0.3.4.0
 synopsis: Serialize to a small byte arrays
 description:
@@ -19,8 +19,8 @@ description:
   of the next chunk. This strategy for building is suitable for most
   CSVs and several line protocols (carbon, InfluxDB, etc.).
   
-homepage: https://github.com/byteverse/small-bytearray-builder
-bug-reports: https://github.com/byteverse/small-bytearray-builder/issues
+homepage: https://github.com/byteverse/bytebuild
+bug-reports: https://github.com/byteverse/bytebuild/issues
 license: BSD-3-Clause
 license-file: LICENSE
 author: Andrew Martin
@@ -36,10 +36,10 @@ flag checked
 
 library
   exposed-modules:
-    Data.ByteArray.Builder
-    Data.ByteArray.Builder.Unsafe
-    Data.ByteArray.Builder.Bounded
-    Data.ByteArray.Builder.Bounded.Unsafe
+    Data.Bytes.Builder
+    Data.Bytes.Builder.Unsafe
+    Data.Bytes.Builder.Bounded
+    Data.Bytes.Builder.Bounded.Unsafe
   reexported-modules:
     Data.Bytes.Chunks
   build-depends:
@@ -72,13 +72,13 @@ test-suite test
   build-depends:
     , QuickCheck >=2.13.1 && <2.14
     , base >=4.12.0.0 && <5
+    , bytebuild
     , byteslice
     , bytestring
     , natural-arithmetic
     , primitive
     , primitive-unlifted >=0.1.2
     , quickcheck-classes >=0.6.4
-    , small-bytearray-builder
     , tasty >=1.2.3 && <1.3
     , tasty-hunit >=0.10.0.2 && <0.11
     , tasty-quickcheck >=0.10.1 && <0.11
@@ -90,10 +90,10 @@ benchmark bench
   type: exitcode-stdio-1.0
   build-depends:
     , base
+    , bytebuild
     , gauge >= 0.2.4
     , natural-arithmetic
     , primitive
-    , small-bytearray-builder
     , text-short
     , byteslice
   ghc-options: -Wall -O2
